Atomize main.go and improve db.Init
